(refactor) : Update footer and header logos, revise address details across multiple components
